Move type extension to context extension
