+++ AutoMapper.AspNetCore.OData.EFCore/OrderBySetting.cs
@@ -1,0 +1,8 @@
+﻿namespace AutoMapper.AspNet.OData
+{
+    internal class OrderBySetting
+    {
+        public string Name { get; set; }
+        public OrderBySetting ThenBy { get; set; }
+    }
+}
